requirements.txt
alpaca-py>=0.33.0
pytz>=2024.1
APScheduler>=3.10.4
pandas_market_calendars>=4.4.0
feedparser>=6.0.11
beautifulsoup4>=4.12.3
requests>=2.32.0
vaderSentiment>=3.3.2
SQLAlchemy>=2.0.30
python-dotenv>=1.0.1
loguru>=0.7.2
pydantic>=2.7.0
pydantic-settings>=2.2.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pdfplumber>=0.11.0
sec-edgar-downloader>=5.0.2
tenacity>=8.3.0
